Refactor: Publish report Step 1 of #1125
